fix(admin): replace all hard-coded api endpoints to the VITE_API_ENDPOINT environment variable
